+++ ext_emconf.php
@@ -8,10 +8,10 @@ $EM_CONF[$_EXTKEY] = [
     'author_email' => 'kevin.lee@wacon.de',
     'state' => 'stable',
     'clearCacheOnLoad' => 0,
-    'version' => '3.4.2',
+    'version' => '4.0.0',
     'constraints' => [
         'depends' => [
-            'typo3' => '12.4.0-13.4.99',
+            'typo3' => '13.4.0-14.0.99',
         ],
         'conflicts' => [],
         'suggests' => [],