+++ CHANGES.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [upgrade-v14-0]
+- [DOCS] Change composer/ext_emconf TYPO3 version to V14
+- [TASK] Migrate Plugin registration/configuration
+- [WIP] Migrate PropertyMapping of selectedAnswers
+
 ## [3.4.2]
 - [BUGFIX] Quiz with single choice could not end due unsufficient support for single and multiple choise answer format